move stores to dedicated file
