+++ frontend/src/components/StoreProfile/StoreProfile.js
@@ -3,7 +3,7 @@ import classNames from "classnames";
 import { Link, withRouter } from "react-router-dom";
 import * as EmailValidator from "email-validator";
 import { URLS } from "../../config";
-import { useParticipantStore } from "../App/App";
+import { useParticipantStore } from "../../util/stores";
 import { shareParticipant} from "../../API";
 import DefaultPage from "../Page/DefaultPage";
 import Loading from "../Loading/Loading";
